no longer create full NP controller
